EZP-25088: rename interfaces
