+++ eZ/Publish/Core/Search/Common/Slot/CreateLocation.php
@@ -12,9 +12,9 @@ namespace eZ\Publish\Core\Search\Common\Slot;
 
 use eZ\Publish\Core\SignalSlot\Signal;
 use eZ\Publish\Core\Search\Common\Slot;
-use eZ\Publish\SPI\Search\Indexer;
-use eZ\Publish\SPI\Search\Indexer\ContentIndexer;
-use eZ\Publish\SPI\Search\Indexer\LocationIndexer;
+use eZ\Publish\SPI\Search\Indexing;
+use eZ\Publish\SPI\Search\Indexing\ContentIndexing;
+use eZ\Publish\SPI\Search\Indexing\LocationIndexing;
 
 /**
  * A Search Engine slot handling CreateLocationSignal.
@@ -32,7 +32,7 @@ class CreateLocation extends Slot
             return;
         }
 
-        if (!$this->searchHandler instanceof Indexer) {
+        if (!$this->searchHandler instanceof Indexing) {
             return;
         }
 
@@ -40,7 +40,7 @@ class CreateLocation extends Slot
             $signal->contentId
         );
 
-        if ($this->searchHandler instanceof ContentIndexer) {
+        if ($this->searchHandler instanceof ContentIndexing) {
             $this->searchHandler->indexContent(
                 $this->persistenceHandler->contentHandler()->load(
                     $signal->contentId,
@@ -49,7 +49,7 @@ class CreateLocation extends Slot
             );
         }
 
-        if ($this->searchHandler instanceof LocationIndexer) {
+        if ($this->searchHandler instanceof LocationIndexing) {
             $this->searchHandler->indexLocation(
                 $this->persistenceHandler->locationHandler()->load($signal->locationId)
             );
