+++ eZ/Publish/Core/Search/Common/Slot/HideLocation.php
@@ -12,8 +12,8 @@ namespace eZ\Publish\Core\Search\Common\Slot;
 
 use eZ\Publish\Core\SignalSlot\Signal;
 use eZ\Publish\Core\Search\Common\Slot;
-use eZ\Publish\SPI\Search\Indexer;
-use eZ\Publish\SPI\Search\Indexer\FullTextIndexer;
+use eZ\Publish\SPI\Search\Indexing;
+use eZ\Publish\SPI\Search\Indexing\FullTextIndexing;
 
 /**
  * A Search Engine slot handling HideLocationSignal.
@@ -31,7 +31,7 @@ class HideLocation extends AbstractSubtree
             return;
         }
 
-        if (!$this->searchHandler instanceof Indexer || $this->searchHandler instanceof FullTextIndexer) {
+        if (!$this->searchHandler instanceof Indexing || $this->searchHandler instanceof FullTextIndexing) {
             return;
         }
