+++ eZ/Publish/Core/Search/Common/Slot/PublishVersion.php
@@ -12,9 +12,9 @@ namespace eZ\Publish\Core\Search\Common\Slot;
 
 use eZ\Publish\Core\SignalSlot\Signal;
 use eZ\Publish\Core\Search\Common\Slot;
-use eZ\Publish\SPI\Search\Indexer\ContentIndexer;
-use eZ\Publish\SPI\Search\Indexer\FullTextIndexer;
-use eZ\Publish\SPI\Search\Indexer\LocationIndexer;
+use eZ\Publish\SPI\Search\Indexing\ContentIndexing;
+use eZ\Publish\SPI\Search\Indexing\FullTextIndexing;
+use eZ\Publish\SPI\Search\Indexing\LocationIndexing;
 
 /**
  * A Search Engine slot handling PublishVersionSignal.
@@ -32,13 +32,13 @@ class PublishVersion extends Slot
             return;
         }
 
-        if ($this->searchHandler instanceof ContentIndexer || $this->searchHandler instanceof FullTextIndexer) {
+        if ($this->searchHandler instanceof ContentIndexing || $this->searchHandler instanceof FullTextIndexing) {
             $this->searchHandler->indexContent(
                 $this->persistenceHandler->contentHandler()->load($signal->contentId, $signal->versionNo)
             );
         }
 
-        if ($this->searchHandler instanceof LocationIndexer) {
+        if ($this->searchHandler instanceof LocationIndexing) {
             $locations = $this->persistenceHandler->locationHandler()->loadLocationsByContent($signal->contentId);
             foreach ($locations as $location) {
                 $this->searchHandler->indexLocation($location);
