+++ eZ/Publish/Core/Search/Common/Slot/Recover.php
@@ -12,10 +12,10 @@ namespace eZ\Publish\Core\Search\Common\Slot;
 
 use eZ\Publish\Core\SignalSlot\Signal;
 use eZ\Publish\Core\Search\Common\Slot;
-use eZ\Publish\SPI\Search\Indexer;
-use eZ\Publish\SPI\Search\Indexer\ContentIndexer;
-use eZ\Publish\SPI\Search\Indexer\FullTextIndexer;
-use eZ\Publish\SPI\Search\Indexer\LocationIndexer;
+use eZ\Publish\SPI\Search\Indexing;
+use eZ\Publish\SPI\Search\Indexing\ContentIndexing;
+use eZ\Publish\SPI\Search\Indexing\FullTextIndexing;
+use eZ\Publish\SPI\Search\Indexing\LocationIndexing;
 
 /**
  * A Search Engine slot handling RecoverSignal.
@@ -33,7 +33,7 @@ class Recover extends Slot
             return;
         }
 
-        if (!$this->searchHandler instanceof Indexer) {
+        if (!$this->searchHandler instanceof Indexing) {
             return;
         }
 
@@ -43,14 +43,14 @@ class Recover extends Slot
         );
 
         foreach ($subtreeIds as $contentId) {
-            if ($this->searchHandler instanceof ContentIndexer || $this->searchHandler instanceof FullTextIndexer) {
+            if ($this->searchHandler instanceof ContentIndexing || $this->searchHandler instanceof FullTextIndexing) {
                 $contentInfo = $contentHandler->loadContentInfo($contentId);
                 $this->searchHandler->indexContent(
                     $contentHandler->load($contentInfo->id, $contentInfo->currentVersionNo)
                 );
             }
 
-            if ($this->searchHandler instanceof LocationIndexer) {
+            if ($this->searchHandler instanceof LocationIndexing) {
                 $this->searchHandler->indexLocation(
                     $this->persistenceHandler->locationHandler()->load($signal->newLocationId)
                 );
