+++ eZ/Publish/Core/Search/Common/Slot/SetContentState.php
@@ -12,10 +12,10 @@ namespace eZ\Publish\Core\Search\Common\Slot;
 
 use eZ\Publish\Core\SignalSlot\Signal;
 use eZ\Publish\Core\Search\Common\Slot;
-use eZ\Publish\SPI\Search\Indexer;
-use eZ\Publish\SPI\Search\Indexer\ContentIndexer;
-use eZ\Publish\SPI\Search\Indexer\FullTextIndexer;
-use eZ\Publish\SPI\Search\Indexer\LocationIndexer;
+use eZ\Publish\SPI\Search\Indexing;
+use eZ\Publish\SPI\Search\Indexing\ContentIndexing;
+use eZ\Publish\SPI\Search\Indexing\FullTextIndexing;
+use eZ\Publish\SPI\Search\Indexing\LocationIndexing;
 
 /**
  * A Search Engine slot handling SetContentStateSignal.
@@ -33,13 +33,13 @@ class SetContentState extends Slot
             return;
         }
 
-        if (!$this->searchHandler instanceof Indexer || $this->searchHandler instanceof FullTextIndexer) {
+        if (!$this->searchHandler instanceof Indexing || $this->searchHandler instanceof FullTextIndexing) {
             return;
         }
 
         $contentInfo = $this->persistenceHandler->contentHandler()->loadContentInfo($signal->contentId);
 
-        if ($this->searchHandler instanceof ContentIndexer) {
+        if ($this->searchHandler instanceof ContentIndexing) {
             $this->searchHandler->indexContent(
                 $this->persistenceHandler->contentHandler()->load(
                     $contentInfo->id,
@@ -48,7 +48,7 @@ class SetContentState extends Slot
             );
         }
 
-        if ($this->searchHandler instanceof LocationIndexer) {
+        if ($this->searchHandler instanceof LocationIndexing) {
             $locations = $this->persistenceHandler->locationHandler()->loadLocationsByContent($contentInfo->id);
             foreach ($locations as $location) {
                 $this->searchHandler->indexLocation($location);
