+++ eZ/Publish/SPI/Search/Indexing/ContentIndexing.php
@@ -4,15 +4,15 @@
  * @copyright Copyright (C) eZ Systems AS. All rights reserved.
  * @license For full copyright and license information view LICENSE file distributed with this source code.
  */
-namespace eZ\Publish\SPI\Search\Indexer;
+namespace eZ\Publish\SPI\Search\Indexing;
 
 use eZ\Publish\SPI\Persistence\Content;
-use eZ\Publish\SPI\Search\Indexer;
+use eZ\Publish\SPI\Search\Indexing;
 
 /**
- * Indexer for handlers that index full text of Content.
+ * Interface for indexing Content items in the search backend.
  */
-interface FullTextIndexer extends Indexer
+interface ContentIndexing extends Indexing
 {
     /**
      * Indexes a content object.
