+++ eZ/Publish/SPI/Search/Indexing/LocationIndexing.php
@@ -4,15 +4,15 @@
  * @copyright Copyright (C) eZ Systems AS. All rights reserved.
  * @license For full copyright and license information view LICENSE file distributed with this source code.
  */
-namespace eZ\Publish\SPI\Search\Indexer;
+namespace eZ\Publish\SPI\Search\Indexing;
 
 use eZ\Publish\SPI\Persistence\Content\Location;
-use eZ\Publish\SPI\Search\Indexer;
+use eZ\Publish\SPI\Search\Indexing;
 
 /**
- * Indexer for handlers that index Locations.
+ * Interface for indexing Locations in the search backend.
  */
-interface LocationIndexer extends Indexer
+interface LocationIndexing extends Indexing
 {
     /**
      * Indexes a Location in the index storage.
